full test refactor
- Now DynamicBackgroundTest are useDynamicColor tests because of the elimination of DynamicBackground and Render components

- Fixed paths for test components to maintain order
