--- src/__test__/components/UseDirection.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-import useDirection from "../../lib/hooks/useDirection";
-
-const UseDirection = () => {
-  const direction = useDirection();
-  return <div>{direction}</div>;
-};
-
-export default UseDirection;
