--- src/__test__/components/UseProximityOnSight.tsx
@@ -1,11 +1,0 @@
-import React, { useRef } from "react";
-import { useProximity } from "../../lib";
-
-const UseProximityOnSight = () => {
-  const ref = useRef(null);
-  const { onSight } = useProximity(ref);
-
-  return <div ref={ref}>{onSight.toString()}</div>;
-};
-
-export default UseProximityOnSight;
